requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pytz==2023.3.post1
requests==2.31.0
pytest==8.2.0
python-dotenv==1.0.1